Reorganize basecoin for better maintainability + add cli bin
