+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "tendermint-basecoin"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 license = "Apache-2.0"
 readme = "README.md"
 categories = ["cryptography::cryptocurrencies"]
@@ -11,6 +11,9 @@ description = """
     An example Tendermint ABCI application that integrates with ibc-rs, making
     use of tendermint-rs.
 """
+
+[[bin]]
+name = "basecoin"
 
 [dependencies]
 base64 = { version = "0.21", default-features = false, features = ["alloc"] }
@@ -26,9 +29,9 @@ serde = "1.0"
 serde_json = "1.0"
 sha2 = "0.10.2"
 structopt = "0.3.21"
-tendermint = "=0.29.0"
-tendermint-abci = "=0.29.0"
-tendermint-proto = "=0.29.0"
+tendermint = "=0.29.1"
+tendermint-abci = "=0.29.1"
+tendermint-proto = "=0.29.1"
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
 tonic = "0.8.3"
 tracing = "0.1.26"
