+++ src/modules/bank/impls.rs
@@ -1,149 +1,27 @@
-use cosmrs::{bank::MsgSend, proto, AccountId, Coin as MsgCoin};
-use displaydoc::Display;
-use ibc_proto::{
-    cosmos::{
-        bank::v1beta1::{
-            query_server::{Query, QueryServer},
-            QueryAllBalancesRequest, QueryAllBalancesResponse, QueryBalanceRequest,
-            QueryBalanceResponse, QueryDenomMetadataRequest, QueryDenomMetadataResponse,
-            QueryDenomOwnersRequest, QueryDenomOwnersResponse, QueryDenomsMetadataRequest,
-            QueryDenomsMetadataResponse, QueryParamsRequest, QueryParamsResponse,
-            QuerySpendableBalancesRequest, QuerySpendableBalancesResponse, QuerySupplyOfRequest,
-            QuerySupplyOfResponse, QueryTotalSupplyRequest, QueryTotalSupplyResponse,
-        },
-        base::v1beta1::Coin as RawCoin,
-    },
-    google::protobuf::Any,
-};
+use super::context::{BankKeeper, BankReader};
+use super::error::Error;
+use super::service::BankService;
+use super::util::{Balances, BalancesPath, Coin, Denom};
+use cosmrs::{bank::MsgSend, proto, AccountId};
+use ibc_proto::{cosmos::bank::v1beta1::query_server::QueryServer, google::protobuf::Any};
 use primitive_types::U256;
 use prost::Message;
-use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, convert::TryInto, fmt::Debug, str::FromStr};
 use tendermint_proto::abci::Event;
-use tonic::{Request, Response, Status};
 use tracing::{debug, trace};
 
-use crate::app::{
-    modules::{
-        auth::{AccountKeeper, AccountReader, AuthAccount, ACCOUNT_PREFIX},
-        Error as ModuleError, Module, QueryResult,
-    },
+use crate::modules::Module;
+use crate::{
+    error::Error as AppError,
+    helper::{Height, Path, QueryResult},
+    modules::auth::account::{AuthAccount, ACCOUNT_PREFIX},
+    modules::auth::context::{AccountKeeper, AccountReader},
     store::{
-        Codec, Height, JsonCodec, JsonStore, Path, ProvableStore, SharedStore, Store, TypedStore,
+        SharedStore,
+        {codec::JsonCodec, Codec},
+        {JsonStore, TypedStore}, {ProvableStore, Store},
     },
 };
-
-#[derive(Debug, Display)]
-pub enum Error {
-    /// failed to decode message
-    MsgDecodeFailure,
-    /// failed to validate message: `{reason}`
-    MsgValidationFailure { reason: String },
-    /// account `{account}` doesn't exist
-    NonExistentAccount { account: AccountId },
-    /// insufficient funds in sender account
-    InsufficientSourceFunds,
-    /// receiver account funds overflow
-    DestFundOverflow,
-    /// Store error: `{reason}`
-    Store { reason: String },
-}
-
-impl From<Error> for ModuleError {
-    fn from(e: Error) -> Self {
-        ModuleError::Bank(e)
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone, Hash, Eq)]
-#[serde(transparent)]
-pub struct Denom(pub String);
-
-#[derive(Clone, Serialize, Deserialize, Debug, Default)]
-pub struct Coin {
-    pub denom: Denom,
-    pub amount: U256,
-}
-
-impl Coin {
-    fn new_empty(denom: Denom) -> Self {
-        Self {
-            denom,
-            amount: 0u64.into(),
-        }
-    }
-}
-
-impl From<(Denom, U256)> for Coin {
-    fn from((denom, amount): (Denom, U256)) -> Self {
-        Self { denom, amount }
-    }
-}
-
-impl From<&MsgCoin> for Coin {
-    fn from(coin: &MsgCoin) -> Self {
-        Self {
-            denom: Denom(coin.denom.to_string()),
-            amount: coin.amount.to_string().parse().unwrap(),
-        }
-    }
-}
-
-/// A mapping of currency denomination identifiers to balances.
-#[derive(Clone, Serialize, Deserialize, Debug, Default)]
-#[serde(transparent)]
-pub struct Balances(Vec<Coin>);
-
-#[derive(Clone, Debug)]
-struct BalancesPath(AccountId);
-
-impl From<BalancesPath> for Path {
-    fn from(path: BalancesPath) -> Self {
-        format!("balances/{}", path.0).try_into().unwrap() // safety - cannot fail as AccountsPath is correct-by-construction
-    }
-}
-
-pub trait BankReader {
-    type Address;
-    type Denom;
-    type Coin;
-    type Coins: IntoIterator<Item = Self::Coin>;
-
-    fn get_all_balances_at_height(&self, height: Height, address: Self::Address) -> Self::Coins;
-
-    fn get_all_balances(&self, address: Self::Address) -> Self::Coins {
-        self.get_all_balances_at_height(Height::Pending, address)
-    }
-}
-
-pub trait BankKeeper {
-    type Error: Debug;
-    type Address: FromStr;
-    type Denom;
-    type Coin;
-
-    /// This function should enable sending ibc fungible tokens from one account to another
-    fn send_coins(
-        &mut self,
-        from: Self::Address,
-        to: Self::Address,
-        amount: impl IntoIterator<Item = Self::Coin>,
-    ) -> Result<(), Self::Error>;
-
-    /// This function to enable minting ibc tokens to a user account
-    fn mint_coins(
-        &mut self,
-        account: Self::Address,
-        amount: impl IntoIterator<Item = Self::Coin>,
-    ) -> Result<(), Self::Error>;
-
-    /// This function should enable burning of minted tokens in a user account
-    fn burn_coins(
-        &mut self,
-        account: Self::Address,
-        amount: impl IntoIterator<Item = Self::Coin>,
-    ) -> Result<(), Self::Error>;
-}
 
 #[derive(Clone)]
 pub struct BankBalanceReader<S> {
@@ -349,9 +227,9 @@ impl<S: 'static + ProvableStore + Default, AR: AccountReader, AK: AccountKeeper>
 }
 
 impl<S: Store, AR: AccountReader, AK: AccountKeeper> Bank<S, AR, AK> {
-    fn decode<T: Message + Default>(message: Any) -> Result<T, ModuleError> {
+    fn decode<T: Message + Default>(message: Any) -> Result<T, AppError> {
         if message.type_url != "/cosmos.bank.v1beta1.MsgSend" {
-            return Err(ModuleError::NotHandled);
+            return Err(AppError::NotHandled);
         }
         Message::decode(message.value.as_ref()).map_err(|_| Error::MsgDecodeFailure.into())
     }
@@ -360,12 +238,12 @@ impl<S: Store, AR: AccountReader, AK: AccountKeeper> Bank<S, AR, AK> {
 impl<S: ProvableStore, AR: AccountReader + Send + Sync, AK: AccountKeeper + Send + Sync> Module
     for Bank<S, AR, AK>
 where
-    <AR as AccountReader>::Address: From<cosmrs::AccountId>,
+    <AR as AccountReader>::Address: From<AccountId>,
     <AK as AccountKeeper>::Account: From<AuthAccount>,
 {
     type Store = S;
 
-    fn deliver(&mut self, message: Any, _signer: &AccountId) -> Result<Vec<Event>, ModuleError> {
+    fn deliver(&mut self, message: Any, _signer: &AccountId) -> Result<Vec<Event>, AppError> {
         let message: MsgSend = Self::decode::<proto::cosmos::bank::v1beta1::MsgSend>(message)?
             .try_into()
             .map_err(|e| Error::MsgValidationFailure {
@@ -414,13 +292,13 @@ where
         _path: Option<&Path>,
         height: Height,
         _prove: bool,
-    ) -> Result<QueryResult, ModuleError> {
+    ) -> Result<QueryResult, AppError> {
         let account_id = match String::from_utf8(data.to_vec()) {
             Ok(s) if s.starts_with(ACCOUNT_PREFIX) => s, // TODO(hu55a1n1): check if valid identifier
-            _ => return Err(ModuleError::NotHandled),
+            _ => return Err(AppError::NotHandled),
         };
 
-        let account_id = AccountId::from_str(&account_id).map_err(|_| ModuleError::NotHandled)?;
+        let account_id = AccountId::from_str(&account_id).map_err(|_| AppError::NotHandled)?;
 
         trace!("Attempting to get account ID: {}", account_id);
 
@@ -448,92 +326,12 @@ where
     fn store(&self) -> &SharedStore<S> {
         &self.store
     }
-}
 
-pub struct BankService<S> {
-    bank_reader: BankBalanceReader<S>,
-}
-
-#[tonic::async_trait]
-impl<S: ProvableStore + 'static> Query for BankService<S> {
-    async fn balance(
-        &self,
-        request: Request<QueryBalanceRequest>,
-    ) -> Result<Response<QueryBalanceResponse>, Status> {
-        debug!("Got bank balance request: {:?}", request);
-
-        let account_id = request
-            .get_ref()
-            .address
-            .parse()
-            .map_err(|e| Status::invalid_argument(format!("{e}")))?;
-        let denom = Denom(request.get_ref().denom.clone());
-        let balances = self.bank_reader.get_all_balances(account_id);
-
-        Ok(Response::new(QueryBalanceResponse {
-            balance: balances
-                .into_iter()
-                .find(|c| c.denom == denom)
-                .map(|coin| RawCoin {
-                    denom: coin.denom.0,
-                    amount: coin.amount.to_string(),
-                }),
-        }))
+    fn check(&self, _message: Any) -> Result<(), AppError> {
+        Ok(())
     }
 
-    async fn all_balances(
-        &self,
-        _request: Request<QueryAllBalancesRequest>,
-    ) -> Result<Response<QueryAllBalancesResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn spendable_balances(
-        &self,
-        _request: Request<QuerySpendableBalancesRequest>,
-    ) -> Result<Response<QuerySpendableBalancesResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn total_supply(
-        &self,
-        _request: Request<QueryTotalSupplyRequest>,
-    ) -> Result<Response<QueryTotalSupplyResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn supply_of(
-        &self,
-        _request: Request<QuerySupplyOfRequest>,
-    ) -> Result<Response<QuerySupplyOfResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn params(
-        &self,
-        _request: Request<QueryParamsRequest>,
-    ) -> Result<Response<QueryParamsResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn denom_metadata(
-        &self,
-        _request: Request<QueryDenomMetadataRequest>,
-    ) -> Result<Response<QueryDenomMetadataResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn denoms_metadata(
-        &self,
-        _request: Request<QueryDenomsMetadataRequest>,
-    ) -> Result<Response<QueryDenomsMetadataResponse>, Status> {
-        unimplemented!()
-    }
-
-    async fn denom_owners(
-        &self,
-        _request: Request<QueryDenomOwnersRequest>,
-    ) -> Result<Response<QueryDenomOwnersResponse>, Status> {
-        unimplemented!()
+    fn begin_block(&mut self, _header: &tendermint::block::Header) -> Vec<Event> {
+        vec![]
     }
 }
