+++ src/modules/bank/mod.rs
@@ -1,0 +1,5 @@
+pub mod context;
+pub mod error;
+pub mod impls;
+pub mod service;
+pub mod util;
