+++ src/modules/mod.rs
@@ -1,0 +1,13 @@
+pub mod auth;
+pub mod bank;
+pub mod ibc;
+pub mod staking;
+
+pub mod traits;
+pub mod types;
+
+pub use self::ibc::{impls::Ibc, transfer::IbcTransferModule};
+pub use auth::impls::Auth;
+pub use bank::impls::Bank;
+pub use staking::impls::Staking;
+pub use traits::{prefix, Identifiable, Module};
