+++ src/store/avl/node.rs
@@ -3,7 +3,7 @@ use std::{borrow::Borrow, mem};
 use sha2::{Digest, Sha256};
 use tendermint::hash::Hash;
 
-use crate::app::store::avl::{as_bytes::AsBytes, proof, HASH_ALGO};
+use crate::store::avl::{as_bytes::AsBytes, proof, HASH_ALGO};
 
 pub type NodeRef<T, V> = Option<Box<AvlNode<T, V>>>;
