+++ src/store/avl/proof.rs
